Fixed data types to avoid potential overflow issues when dealing with strings

+++ string/ft_strcpy.c
@@ -6,7 +6,7 @@
 /*   By: chlimous <chlimous@student.42.fr>	    +#+  +:+	   +#+	      */
 /*						  +#+#+#+#+#+	+#+	      */
 /*   Created: 2023/09/07 00:33:41 by chlimous	       #+#    #+#	      */
-/*   Updated: 2023/12/31 04:51:02 by chlimous         ###   ########.fr       */
+/*   Updated: 2024/01/05 23:46:00 by chlimous         ###   ########.fr       */
 /*									      */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strcpy(char *dest, const char *src)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (src[i] != '\0')
